refactor(form,detail): tách làm 2 file để quản lý boc
